requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
python-dotenv==1.0.1
PyJWT==2.10.1
passlib[bcrypt]==1.7.4
bcrypt==4.1.3  # Pin: bcrypt 5.0+ incompatible with passlib
pydantic==2.10.3
email-validator==2.2.0
pydantic-settings==2.6.1
